refactor(capabilities): migrate the ui cap to the struct-hosted actor macro
Move the `impl NativeActor for UiCapability` block from `ui/widgets.rs` into
`ui/runtime.rs` under `#[runtime]` and lift `#[actor(singleton)]` onto the
struct (#2383).
